ml-service/requirements.txt
tensorflow
keras
scikit-learn
pandas
numpy
matplotlib
seaborn
lightgbm
flask
flask-cors
python-dotenv
nltk
